add script to process ledger objects
