+++ docs/source/xrpl.models.ledger_objects.rst
@@ -1,0 +1,21 @@
+xrpl.models.ledger\_objects package
+===================================
+
+Submodules
+----------
+
+xrpl.models.ledger\_objects.ledger\_entry\_type module
+------------------------------------------------------
+
+.. automodule:: xrpl.models.ledger_objects.ledger_entry_type
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+Module contents
+---------------
+
+.. automodule:: xrpl.models.ledger_objects
+   :members:
+   :undoc-members:
+   :show-inheritance:
