+++ xrpl/models/ledger_objects/ledger_entry_type.py
@@ -1,0 +1,7 @@
+"""The different ledger entry types"""
+
+from enum import Enum
+
+
+class LedgerEntryType(str, Enum):
+    """The different ledger entry types"""
